Updated the time of the observation and the points in a period.
On the current setup, run:
python waterfaller_puma.py -T 0.0 -t 84.0 -d 68.0 ds16384_J0835-4510_A2_20191124_063438.fil
and you will get 938 plots of different pulses for 1.4 min of observation and
one cvs file with one pulse per row and 205 points per pulse (205 columns).

In this observation, there are 305.5 points per period (89.4 ms). We altern
one 305 and one 306. This works for this amount of pulses but the peak shifts
to the right. We have to do something else for more time of observation.

We have to modify the range and the interval printed in the csv accordingly
with the observation under consideration. Also the flag -t on the terminal.