Generates the individual pulses
This script takes as imput the .csv files generated with the waterfall script.
Each .csv file has data for a time = pulse_period * 1000.
In this script all this data is unified in an array that is interpolated to
obtain the points we want in order to have a perfect cut in period.
The output of this script is a .csv with a pulse in each line (you need to put
by hand how many pulses you want to obtain and how many points per pulse).